Allow pointer receiver methods to be used (issue#21)
