+++ .travis.yml
@@ -4,3 +4,7 @@ install: go get -t -v ./...
 go:
     - 1.7
     - 1.8
+    - 1.9
+    - 1.10
+    - 1.11
+    - 1.12
